+++ RELEASENOTES.md
@@ -3,6 +3,7 @@
 ### Unreleased changes
 
 *   Common Library:
+    *   Replace `SimpleBasePlayer.State.playlist` by `getPlaylist()` method.
 *   ExoPlayer:
     *   `MediaCodecRenderer.onProcessedStreamChange()` can now be called for
         every media item. Previously it was not called for the first one. Use